minor update - core.Command
